[RFC] Remove build phases
As discussed today between @matz3, @tommyvinhlam, @codeworrior, @randombyte

Build phases would allow for more precise extensibility of a types build
execution. Mainly scenarios where only a single task should be replaced
or added could overwrite one of the build phases to hook in-between of
certain sets of tasks.

Instead, they will have to overwrite the whole build function containing
all tasks. The main reason for this decision is the lack of known
extensibility use-cases. Therefore we do not want to decide on the
naming and semantics of any build phases yet as changing them later on
will hardly be possible.