Remove getSpecVersion() from Project interface
With the latest changes in https://github.com/SAP/ui5-project/pull/431,
the method now returns an instance of a "SpecificationVersion" class.
This would require an additional specVersion-dependent interface.

Since use cases in custom tasks and -middelare are not clear, we should
rather keep it simple for now and do not provide this method.